+++ capstone/yieldpay/programs/yieldpay/src/state/config.rs
@@ -13,7 +13,7 @@ pub struct Config {
     pub total_users: u64,
     pub total_merchants: u64,
     pub yield_min_period: u64,
-    pub apy_bps: u8,
+    pub apy_bps: u64,
     pub config_bump: u8,
     pub yield_bump: u8,
 }
